refactor: utils.js를 ts로 변경 완료
